+++ openshift-client/src/test/java/io/fabric8/openshift/client/internal/OpenShiftOAuthInterceptorTest.java
@@ -19,6 +19,7 @@ import io.fabric8.kubernetes.client.Config;
 import io.fabric8.kubernetes.client.http.HttpClient;
 import io.fabric8.kubernetes.client.http.StandardHttpRequest;
 import io.fabric8.kubernetes.client.http.TestHttpResponse;
+import io.fabric8.kubernetes.client.http.WebSocket;
 import io.fabric8.kubernetes.client.utils.TokenRefreshInterceptor;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
@@ -71,7 +72,7 @@ class OpenShiftOAuthInterceptorTest {
   }
 
   @Test
-  void testTokenRefreshedFromConfig() {
+  void testTokenRefreshedFromConfigForHttpRequestBuilder() {
     Config config = Mockito.mock(Config.class, RETURNS_DEEP_STUBS);
     Mockito.when(config.refresh().getOauthToken()).thenReturn("token");
 
@@ -83,9 +84,26 @@ class OpenShiftOAuthInterceptorTest {
     builder.uri("http://localhost");
 
     interceptor.afterFailure(builder, TestHttpResponse.from(HttpURLConnection.HTTP_UNAUTHORIZED, "not for you")
-        .withRequest(new StandardHttpRequest(null, URI.create("http://localhost"), "GET", null)), null);
+      .withRequest(new StandardHttpRequest(null, URI.create("http://localhost"), "GET", null)), null);
 
     assertEquals(Collections.singletonList("Bearer token"), builder.build().headers(TokenRefreshInterceptor.AUTHORIZATION));
+    Mockito.verify(config).setOauthToken("token");
+  }
+
+  @Test
+  void testTokenRefreshedFromConfigForWebSocketBuilder() {
+    Config config = Mockito.mock(Config.class, RETURNS_DEEP_STUBS);
+    Mockito.when(config.refresh().getOauthToken()).thenReturn("token");
+
+    HttpClient client = Mockito.mock(HttpClient.class);
+    OpenShiftOAuthInterceptor interceptor = new OpenShiftOAuthInterceptor(client, config);
+
+    WebSocket.Builder builder = Mockito.mock(WebSocket.Builder.class, RETURNS_DEEP_STUBS);
+
+    interceptor.afterFailure(builder, TestHttpResponse.from(HttpURLConnection.HTTP_UNAUTHORIZED, "not for you")
+      .withRequest(new StandardHttpRequest(null, URI.create("http://localhost"), "GET", null)), null);
+
+    Mockito.verify(builder).setHeader(TokenRefreshInterceptor.AUTHORIZATION, "Bearer token");
     Mockito.verify(config).setOauthToken("token");
   }